+++ src/test/java/dev/vianneynara/portoapi/PortoApiApplicationTests.java
@@ -1,4 +1,4 @@
-package dev.vianneynara.simpleportoapi;
+package dev.vianneynara.portoapi;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
